requirements.txt
# BeagleBone Blue Robot Control Server Dependencies
# For bbb-server.py (runs on BBB)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
psutil>=5.9.0

# For dashboard.py (runs on MacBook)
paramiko>=3.4.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
